Separate mesos master and slave options
There is a conflict between mesos::property resource names when
mesos::master::options and mesos::slave::options are used on same
machine. The file is different since the directory is different but
the resource name is just the key. There is also a potential conflict
between attributes and resources names.

The solution is to add prefix to name like is done for the static
properties. This involves adding parameter to mesos_hash_parser
function.